refactor: logic separation
