+++ server/README.md
@@ -1,6 +1,6 @@
 # Flask Server
 
-This is a Flask application for prediction.
+This is a Flask application for cnn prediction.
 
 ## Getting Started
 
@@ -12,7 +12,7 @@ Follow these instructions to set up the project on your local machine.
 
 ### Setting Up a Virtual Environment
 
-It's recommended to create a virtual environment to isolate project dependencies. Open a terminal navigate to the root of the project and run the following commands:
+It's recommended to create a virtual environment to isolate project dependencies. Open a terminal navigate to the root directory of the project and run the following commands:
 
 ```bash
 # Create a virtual environment (replace 'myenv' with your preferred name)
@@ -28,5 +28,18 @@ Once you have activated the virtual environment, you can install the project dep
 ```bash
 pip install -r requirements.txt
 ```
+
+If you install anyother dependencies make sure to update the requirements.txt by running the following command
+
+```bash
+pip freeze > requirements.txt
+```
+
+### Run app
+```bash
+python main.py
+```
+
+
