+++ server/app/__init__.py
@@ -1,0 +1,7 @@
+# app/__init__.py
+from flask import Flask
+
+app = Flask(__name__)
+
+# Import routes
+from app import routes
