+++ server/app/controllers/__init_.py
@@ -1,0 +1,1 @@
+# app/controllers/__init__.py
