+++ server/main.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from app import app  # Import the 'app' instance from your 'app' package
+
+if __name__ == '__main__':
+    app.run(debug=True)
+
